completed hw7

+++ hw7/Renderer.hpp
@@ -6,7 +6,7 @@
 #pragma once
 struct hit_payload
 {
-    float tNear;
+    double tNear;
     uint32_t index;
     Vector2f uv;
     Object* hit_obj;
