+++ hw7/Vector.hpp
@@ -11,16 +11,16 @@
 
 class Vector3f {
 public:
-    float x, y, z;
+    double x, y, z;
     Vector3f() : x(0), y(0), z(0) {}
-    Vector3f(float xx) : x(xx), y(xx), z(xx) {}
-    Vector3f(float xx, float yy, float zz) : x(xx), y(yy), z(zz) {}
-    Vector3f operator * (const float &r) const { return Vector3f(x * r, y * r, z * r); }
-    Vector3f operator / (const float &r) const { return Vector3f(x / r, y / r, z / r); }
+    Vector3f(double xx) : x(xx), y(xx), z(xx) {}
+    Vector3f(double xx, double yy, double zz) : x(xx), y(yy), z(zz) {}
+    Vector3f operator * (const double &r) const { return Vector3f(x * r, y * r, z * r); }
+    Vector3f operator / (const double &r) const { return Vector3f(x / r, y / r, z / r); }
 
-    float norm() {return std::sqrt(x * x + y * y + z * z);}
+    double norm() {return std::sqrt(x * x + y * y + z * z);}
     Vector3f normalized() {
-        float n = std::sqrt(x * x + y * y + z * z);
+        double n = std::sqrt(x * x + y * y + z * z);
         return Vector3f(x / n, y / n, z / n);
     }
 
@@ -29,7 +29,7 @@ public:
     Vector3f operator + (const Vector3f &v) const { return Vector3f(x + v.x, y + v.y, z + v.z); }
     Vector3f operator - () const { return Vector3f(-x, -y, -z); }
     Vector3f& operator += (const Vector3f &v) { x += v.x, y += v.y, z += v.z; return *this; }
-    friend Vector3f operator * (const float &r, const Vector3f &v)
+    friend Vector3f operator * (const double &r, const Vector3f &v)
     { return Vector3f(v.x * r, v.y * r, v.z * r); }
     friend std::ostream & operator << (std::ostream &os, const Vector3f &v)
     { return os << v.x << ", " << v.y << ", " << v.z; }
@@ -56,28 +56,28 @@ class Vector2f
 {
 public:
     Vector2f() : x(0), y(0) {}
-    Vector2f(float xx) : x(xx), y(xx) {}
-    Vector2f(float xx, float yy) : x(xx), y(yy) {}
-    Vector2f operator * (const float &r) const { return Vector2f(x * r, y * r); }
+    Vector2f(double xx) : x(xx), y(xx) {}
+    Vector2f(double xx, double yy) : x(xx), y(yy) {}
+    Vector2f operator * (const double &r) const { return Vector2f(x * r, y * r); }
     Vector2f operator + (const Vector2f &v) const { return Vector2f(x + v.x, y + v.y); }
-    float x, y;
+    double x, y;
 };
 
-inline Vector3f lerp(const Vector3f &a, const Vector3f& b, const float &t)
+inline Vector3f lerp(const Vector3f &a, const Vector3f& b, const double &t)
 { return a * (1 - t) + b * t; }
 
 inline Vector3f normalize(const Vector3f &v)
 {
-    float mag2 = v.x * v.x + v.y * v.y + v.z * v.z;
+    double mag2 = v.x * v.x + v.y * v.y + v.z * v.z;
     if (mag2 > 0) {
-        float invMag = 1 / sqrtf(mag2);
+        double invMag = 1 / sqrtf(mag2);
         return Vector3f(v.x * invMag, v.y * invMag, v.z * invMag);
     }
 
     return v;
 }
 
-inline float dotProduct(const Vector3f &a, const Vector3f &b)
+inline double dotProduct(const Vector3f &a, const Vector3f &b)
 { return a.x * b.x + a.y * b.y + a.z * b.z; }
 
 inline Vector3f crossProduct(const Vector3f &a, const Vector3f &b)
